youtube_summarizer/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
streamlit==1.35.0
youtube_transcript_api==0.6.2
google-generativeai==0.4.0
